+++ covasim/interventions.py
@@ -81,6 +81,8 @@ class Intervention:
     dict format, which they can be recreated from. To display all the attributes
     of the intervention, use disp() instead.
 
+    To retrieve a particular intervention from a sim, use sim.get_intervention().
+
     Args:
         label (str): a label for the intervention (used for plotting, and for ease of identification)
         show_label (bool): whether or not to include the label, if provided, in the legend
@@ -141,14 +143,13 @@ class Intervention:
 
     def apply(self, sim):
         '''
-        Apply intervention
-
-        Function signature matches existing intervention definition
-        This method gets called at each timestep and must be implemented
-        by derived classes
+        Apply the intervention. This is the core method which each drived intervention
+        class must implement. This method gets called at each timestep and can make
+        arbitrary changes to the Sim object, as well as storing or modifying the
+        state of the intervention.
 
         Args:
-            sim: The Sim instance
+            sim: the Sim instance
 
         Returns:
             None
@@ -158,7 +159,7 @@ class Intervention:
 
     def plot(self, sim, ax=None, **kwargs):
         '''
-        Call function during plotting
+        Plot the intervention
 
         This can be used to do things like add vertical lines on days when
         interventions take place. Can be disabled by setting self.do_plot=False.
@@ -193,7 +194,10 @@ class Intervention:
         one-way export to produce a JSON-compatible representation of the
         intervention. In the first instance, the object dict will be returned.
         However, if an intervention itself contains non-standard variables as
-        attributes, then its `to_json` method will need to handle those
+        attributes, then its `to_json` method will need to handle those.
+
+        Note that simply printing an intervention will usually return a representation
+        that can be used to recreate it.
 
         Returns:
             JSON-serializable representation (typically a dict, but could be anything else)
@@ -568,9 +572,9 @@ def get_quar_inds(quar_policy, sim):
     '''
     t = sim.t
     if   quar_policy is None:    quar_test_inds = np.array([])
-    elif quar_policy == 'start': quar_test_inds = cvu.true(sim.people.date_quarantined==t-1) # Actually do the day before since testing usually happens before contact tracing
-    elif quar_policy == 'end':   quar_test_inds = cvu.true(sim.people.date_end_quarantine==t)
-    elif quar_policy == 'both':  quar_test_inds = np.concatenate([cvu.true(sim.people.date_quarantined==t-1), cvu.true(sim.people.date_end_quarantine==t)])
+    elif quar_policy == 'start': quar_test_inds = cvu.true(sim.people.date_quarantined==t-1) # Actually do the day after since testing usually happens before contact tracing
+    elif quar_policy == 'end':   quar_test_inds = cvu.true(sim.people.date_end_quarantine==t+1) # +1 since they are released on date_end_quarantine, so do the day before
+    elif quar_policy == 'both':  quar_test_inds = np.concatenate([cvu.true(sim.people.date_quarantined==t-1), cvu.true(sim.people.date_end_quarantine==t+1)])
     elif quar_policy == 'daily': quar_test_inds = cvu.true(sim.people.quarantined)
     elif sc.isnumber(quar_policy) or (sc.isiterable(quar_policy) and not sc.isstring(quar_policy)):
         quar_policy = sc.promotetoarray(quar_policy)