+++ CHANGELOG.rst
@@ -16,14 +16,14 @@ Latest versions (1.7.x)
 
 Version 1.7.1 (2020-09-23)
 --------------------------
-- Added new methods, ``sim.get_interventions()`` and ``sim.get_analyzers()``, which return interventions or analyzers based on the index, label, or type.
+- Added two new methods, ``sim.get_interventions()`` and ``sim.get_analyzers()``, which return interventions or analyzers based on the index, label, or type.
 - Added a new analyzer, ``cv.daily_stats()``, which can print out and plot detailed information about the state of the simulation on each day.
 - MultiSims can now be run without parallelization; use ``msim.run(parallel=False)``. This can be useful for debugging, or for parallelizing across rather than within MultiSims (since ``multiprocessing`` calls cannot be nested).
 - ``sim.people.not_defined()`` has been renamed ``sim.people.undefined()``.
 - New helper functions have been added: ``cv.maximize()`` maximizes the current figure, and ``cv.get_rows_cols()`` converts a number (usually a number of plots) into the required number of rows and columns. Both will eventually be moved to Sciris.
-- Corrected the transmission tree plot to account for people who have left quarantine.
+- The transmission tree plot has been corrected to account for people who have left quarantine. The definition of "quarantine end" for the sake of testing (``quar_policy='end'`` for ``cv.test_num()`` and ``cv.test_prob()``) has also been shifted up by a day (since by ``date_end_quarantine``, people are no longer in quarantine by the end of the day, so tests were not being counted as happening in quarantine).
 - Code has been moved between ``sim.py``, ``people.py``, and ``base.py`` to better reflect the division between "the simulation" (the first two files) and "the housekeeping" (the last file).
-- *Regression info*: Changes to quarantine will mean that simulations that used contact tracing will show different results (although since recovered people do not transmit, differences should mostly just be stochastic). For backwards incompatibility, only user scripts that explicitly call ``sim.people.not_defined()`` should require updating.
+- *Regression info*: Scripts that used ``quar_policy='end'`` may now provide stochastically different results. User scripts that explicitly call ``sim.people.not_defined()`` should be updated to call ``sim.people.undefined()`` instead.
 - *GitHub info*: PR `690 <https://github.com/amath-idm/covasim/pull/690>`__